Flatten the Store hierarchy
Only one implementation now: prefer simplicity.
